Create a helper method to reduce lines of code
