+++ spec/rails_helper.rb
@@ -11,6 +11,7 @@ require 'capybara/rspec'
 require 'support/factorybot'
 require_relative 'support/controller_macros'
 Dir[Rails.root.join('spec', 'support', '**', '*.rb')].sort.each { |f| require f }
+Dir[Rails.root.join('spec', 'helpers', '**', '*.rb')].sort.each { |f| require f }
 
 begin
   ActiveRecord::Migration.maintain_test_schema!
